+++ facebook_auth_desktop/example/lib/main.dart
@@ -15,7 +15,7 @@ void main() {
   plugin.webAndDesktopInitialize(
     appId: '1329834907365798',
     cookie: true,
-    version: 'v15.0',
+    version: 'v18.0',
     xfbml: true,
   );
   runApp(const MyApp());
@@ -32,8 +32,19 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class HomePage extends StatelessWidget {
+class HomePage extends StatefulWidget {
   const HomePage({Key? key}) : super(key: key);
+
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  void initState() {
+    super.initState();
+    // plugin.accessToken.then((value) => null);
+  }
 
   void _signIn(BuildContext context) async {
     final result = await plugin.login();