+++ facebook_auth/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 6.0.3
+- Updated to `facebook_auth_desktop: ^1.0.2`
+- Updated Facebook iOS SDK to 16.2.0
+- Updated Facebook Android SDK to 16.2.0
+
 ### 6.0.2
 - Updated to `facebook_auth_desktop: ^1.0.1`